# feat: Optimise `sortBreadcrumbs()`